+++ testWrapper/tests/FirstOrder_test.cpp
@@ -7,15 +7,15 @@ NTFX_TEST_BEGIN
 
 NTFX_TEST() {
   auto firstOrderHpf =
-      NtFx::FirstOrder::StereoFilter<float, NtFx::FirstOrder::Shape::hpf>();
+      NtFx::FirstOrder::StereoFilter<double, NtFx::FirstOrder::Shape::hpf>();
   firstOrderHpf.setFc(1e3);
   NTFX_ADD_TEST(firstOrderHpf, "impulse");
   auto firstOrderLpf =
-      NtFx::FirstOrder::StereoFilter<float, NtFx::FirstOrder::Shape::lpf>();
+      NtFx::FirstOrder::StereoFilter<double, NtFx::FirstOrder::Shape::lpf>();
   firstOrderLpf.setFc(1e3);
   NTFX_ADD_TEST(firstOrderLpf, "impulse");
-  auto firstOrderLpfWithZero =
-      NtFx::FirstOrder::StereoFilter<float, NtFx::FirstOrder::Shape::lpfZero>();
+  auto firstOrderLpfWithZero = NtFx::FirstOrder::StereoFilter<double,
+      NtFx::FirstOrder::Shape::lpfZero>();
   firstOrderLpfWithZero.setFc(1e3);
   NTFX_ADD_TEST(firstOrderLpfWithZero, { "impulse" });
   return NTFX_RUN_TESTS();
