+++ testWrapper/tests/PeakSensor_test.cpp
@@ -6,10 +6,10 @@
 NTFX_TEST_BEGIN
 
 NTFX_TEST() {
-  auto peakSensor = NtFx::PeakSensorStereo<float>();
+  auto peakSensor = NtFx::PeakSensorStereo<double>();
   peakSensor.setT_ms(20);
   NTFX_ADD_TEST(peakSensor, "dynamic");
-  auto peakHoldSensor = NtFx::PeakHoldSensorStereo<float>();
+  auto peakHoldSensor = NtFx::PeakHoldSensorStereo<double>();
   peakHoldSensor.setT_ms(20);
   peakHoldSensor.setTHold_ms(1);
   NTFX_ADD_TEST(peakHoldSensor, "dynamic");
